+++ apps/web/src/app/(needLogin)/mypage/badge/_components/HowToNotification/HowToNotification.types.ts
@@ -1,0 +1,8 @@
+type infoType = {
+  title: string;
+  description: string;
+};
+
+export type infoProps = {
+  info: infoType;
+};
