+++ apps/web/src/app/(needLogin)/mypage/badge/_components/OneBadgeSlide/OneBadgeSlide.types.tsx
@@ -1,0 +1,20 @@
+import { StaticImageData } from 'next/image';
+
+type ItemType = {
+  title: string;
+  slug: string;
+  description: {
+    title: string;
+    subtitle?: string;
+  }[];
+  info: {
+    title: string;
+    description: string;
+  };
+};
+
+export type OneBadgeSlideProps = {
+  item: ItemType;
+  badgeDefault: StaticImageData;
+  id: number;
+};
