+++ apps/web/src/app/(needLogin)/mypage/setting/service/_components/OneInquiryHistory/OneInquiryHistory.tsx
@@ -3,14 +3,7 @@ import { Button } from 'ui';
 
 import { isIndividualToggleAtom } from '@/store';
 
-import { InquiryMockData } from '../../_mock/InquiryMocks';
-
-// 목데이터이므로 임시 type 설정 -> 추후 실제 데이터 타입으로 변경
-type InquiryData = (typeof InquiryMockData)[0];
-
-type InquiryDataProps = {
-  data: InquiryData;
-};
+import type { InquiryDataProps } from './OneInquiryHistory.types';
 
 export const OneInquiryHistory = ({ data }: InquiryDataProps) => {
   const [isIndividualToggle, setIsIndividualToggle] = useAtom(
