+++ apps/web/src/app/(needLogin)/mypage/setting/service/_components/OneInquiryHistory/OneInquiryHistory.types.tsx
@@ -1,0 +1,7 @@
+import { InquiryMockData } from '../../_mock/InquiryMocks';
+
+type InquiryData = (typeof InquiryMockData)[0];
+
+export type InquiryDataProps = {
+  data: InquiryData;
+};
