+++ apps/web/src/components/Profile/_components/FlagInfo/FlagInfo.types.tsx
@@ -1,0 +1,16 @@
+import { StaticImageData } from 'next/image';
+
+type DataType = {
+  ko: string;
+  en: string;
+  bannerImg: StaticImageData;
+  continentData: {
+    img: StaticImageData;
+    name: string;
+  }[];
+};
+
+export type FlagInfoDataProps = {
+  data: DataType;
+  id: number;
+};
