+++ apps/web/src/components/Profile/_components/OneFlagInfo/OneFlagInfo.types.tsx
@@ -1,0 +1,10 @@
+import { StaticImageData } from 'next/image';
+
+type OneFlagInfoType = {
+  img: StaticImageData;
+  name: string;
+};
+
+export type OneFlagInfoProps = {
+  item: OneFlagInfoType;
+};
